+++ src/main/java/eu/iu3cyv/status/monitor/CheckResultRepository.java
@@ -1,0 +1,6 @@
+package eu.iu3cyv.status.monitor;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CheckResultRepository extends JpaRepository<CheckResult, Long> {
+}