+++ src/main/java/eu/iu3cyv/status/monitor/MonitorRepository.java
@@ -7,4 +7,6 @@ import java.util.List;
 public interface MonitorRepository extends JpaRepository<Monitor, Long> {
 
     List<Monitor> findByEnabledTrueOrderByDisplayOrderAscNameAsc();
+
+    List<Monitor> findAllByOrderByDisplayOrderAscNameAsc();
 }